+++ app/src/main/java/com/mmdev/roove/ui/settings/edit/SettingsEditInfoPhotoAdapter.kt
@@ -1,14 +1,14 @@
 /*
  * Created by Andrii Kovalchuk
  * Copyright (c) 2020. All rights reserved.
- * Last modified 01.03.20 19:02
+ * Last modified 02.03.20 16:11
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package com.mmdev.roove.ui.settings
+package com.mmdev.roove.ui.settings.edit
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
